+++ src/core/message.ts
@@ -9,6 +9,7 @@ import { AnySignatureType, Signature, SignatureType } from './signature';
 export interface MsgReceipt<T extends object> {
   get result(): Nillable<T[]>;
   logs?: string;
+  error?: string;
 }
 
 /**